refactoring: use strict integer types

+++ cpio.h
@@ -8,7 +8,7 @@ typedef void(*putchar_cb_t)(char);
 
 void cpio_read_catalog(putchar_cb_t putchar_cb);
 int cpio_file_size(const char *fname);
-int cpio_read_file(const char *fname, char *buffer, int buffer_size);
+int cpio_read_file(const char *fname, char *buffer, size_t buffer_size);
 
 void cpio_set_initrd_start(uint64_t addr);
